receiving all nodes battery stats, need to define the nodes to make a measure

+++ ApplicationClusteringImplementations/src/BatteryStatsInfo/SimpleBatteryStatsInfo.cc
@@ -24,7 +24,7 @@ SimpleBatteryStatsInfo::~SimpleBatteryStatsInfo()
 }
 
 
-SimpleBatteryStatsInfo::SimpleBatteryStatsInfo(const char *name, int residualRelative, bool namepooling) : cNamedObject(name, namepooling)
+SimpleBatteryStatsInfo::SimpleBatteryStatsInfo(const char *name, double residualRelative, bool namepooling) : cNamedObject(name, namepooling)
 {
     this->residualRelative = residualRelative;
 }
